Skip decrypting GFWL save data When the game loads a save file it will check the file's CRC values against the XLive decrypted buffer. However, XLive will not decrypt a protected buffer if it does not match the currently connected GFWL account. We can simply skip the decryption step and instead write the required CRC values into the unprotected buffer.
This enables us to share save files with other (local) accounts.
Obviously this is a small security concern if one tries to load an
external save file. Make sure to always trust your source :^)

This is also the first step for a custom save manager and the first step
to get rid of GFWL :^)